requirements.txt
langchain-groq
langchain
chainlit
pydantic
openai
python-dotenv
google-generativeai